Refactor: 변경된 컨벤션에 맞도록 코드 수정 - Controller: Put -> Patch - 객체가 존재하는 ResponseEntity 수정 - 일부 응답 인라인화
