+++ src/main/java/com/erp/domain/client/service/ClientService.java
@@ -3,10 +3,10 @@ package com.erp.domain.client.service;
 import com.erp.domain.accident.entity.Accident;
 import com.erp.domain.accident.repository.AccidentRepository;
 import com.erp.domain.client.dto.request.AddClientBlacklist;
-import com.erp.domain.client.dto.responce.ClientAccidentHistoryResponse;
-import com.erp.domain.client.dto.responce.ClientDetailResponse;
-import com.erp.domain.client.dto.responce.ClientRentHistoryResponse;
-import com.erp.domain.client.dto.responce.ClientSummaryResponse;
+import com.erp.domain.client.dto.response.ClientAccidentHistoryResponse;
+import com.erp.domain.client.dto.response.ClientDetailResponse;
+import com.erp.domain.client.dto.response.ClientRentHistoryResponse;
+import com.erp.domain.client.dto.response.ClientSummaryResponse;
 import com.erp.domain.client.entity.Client;
 import com.erp.domain.client.repository.ClientRepository;
 import com.erp.domain.rent.entity.Rent;
